cleanup code

+++ src/main/java/com/example/hotelsapisoap/service/HotelServiceImpl.java
@@ -1,23 +1,23 @@
 package com.example.hotelsapisoap.service;
 
+import com.example.hotelsapisoap.exception.BadRequestException;
+import com.example.hotelsapisoap.exception.NotFoundException;
 import com.example.hotelsapisoap.model.Amenity;
 import com.example.hotelsapisoap.model.Hotel;
 import com.example.hotelsapisoap.repository.AmenityRepository;
 import com.example.hotelsapisoap.repository.HotelRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import java.util.Optional;
 
 @Service
 public class HotelServiceImpl implements HotelService {
-    private HotelRepository repository;
-    private AmenityRepository amenityRepository;
+    private final HotelRepository repository;
+    private final AmenityRepository amenityRepository;
 
-    @Autowired
     public HotelServiceImpl(HotelRepository repository, AmenityRepository amenityRepository){
         this.repository = repository;
         this.amenityRepository = amenityRepository;
@@ -25,11 +25,16 @@ public class HotelServiceImpl implements HotelService {
 
     @Override
     public Hotel getById(long id){
-        return repository.findById(id);
+        Optional<Hotel> hotel = repository.findById(id);
+
+        if(!hotel.isPresent())
+            throw new NotFoundException("Hotel with id: " + id + " not found.");
+
+        return hotel.get();
     };
 
     @Override
-    public Page<Hotel> filterByName(int page, String name) {
+    public Page<Hotel> getHotels(int page, String name) {
         Pageable pageable = PageRequest.of(page, 5);
 
         if(!name.equals("")) {
@@ -39,46 +44,78 @@ public class HotelServiceImpl implements HotelService {
         return repository.findAll(pageable);
     }
 
-    public Hotel save(Hotel hotel) {
-        try {
-            return repository.save(hotel);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return null;
-        }
+    @Override
+    public Hotel addHotel(Hotel hotel) throws BadRequestException {
+        Optional<Hotel> optionalHotel = repository.findById(hotel.getId());
+        if(optionalHotel.isPresent())
+            throw new BadRequestException("Hotel with id: " + hotel.getId() + " already exists");
+
+        if(hotel.getRating() < 1 || hotel.getRating() > 5)
+            throw new BadRequestException("Invalid value for rating. Provide an integer value from 1 to 5;");
+
+        return repository.save(hotel);
     }
 
+    @Override
+    public Hotel editHotel(Hotel hotel) throws NotFoundException {
+        Optional<Hotel> optionalHotel = repository.findById(hotel.getId());
+
+        if(!optionalHotel.isPresent())
+            throw new NotFoundException("Hotel with id: " + hotel.getId() + " not found");
+
+        if(hotel.getRating() < 1 || hotel.getRating() > 5)
+            throw new BadRequestException("Invalid value for rating. Provide an integer value from 1 to 5;");
+
+        Hotel hotelToSave = optionalHotel.get();
+        hotelToSave.setName(hotel.getName());
+        hotelToSave.setAddress(hotel.getAddress());
+        hotelToSave.setRating(hotel.getRating());
+        return repository.save(hotelToSave);
+    }
+
+    @Override
     public Hotel addAmenity(long id_hotel, long id_amenity) {
-        try {
-            Hotel hotel = repository.findById(id_hotel);
+        Optional<Hotel> optionalHotel = repository.findById(id_hotel);
 
-            Amenity amenity = amenityRepository.findById(id_amenity);
+        if(!optionalHotel.isPresent())
+            throw new NotFoundException("Hotel with id: " + id_hotel + " not found");
 
-            hotel.addAmenity(amenity);
-            return repository.save(hotel);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return null;
-        }
+        Optional<Amenity> optionalAmenity = amenityRepository.findById(id_amenity);
+
+        if(!optionalAmenity.isPresent())
+            throw new NotFoundException("Amenity with id: " + id_amenity + " not found");
+
+        Hotel hotel = optionalHotel.get();
+        Amenity amenity = optionalAmenity.get();
+
+        hotel.addAmenity(amenity);
+
+        return repository.save(hotel);
     }
 
     public Hotel removeAmenity(long id_hotel, long id_amenity) {
-        try {
-            Hotel hotel = repository.findById(id_hotel);
-            hotel.removeAmenity(id_amenity);
-            return repository.save(hotel);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return null;
-        }
+        Optional<Hotel> optionalHotel = repository.findById(id_hotel);
+
+        if(!optionalHotel.isPresent())
+            throw new NotFoundException("Hotel with id: " + id_hotel + " not found");
+
+        Optional<Amenity> optionalAmenity = amenityRepository.findById(id_amenity);
+
+        if(!optionalAmenity.isPresent())
+            throw new NotFoundException("Amenity with id: " + id_amenity + " not found");
+
+        Hotel hotel = optionalHotel.get();
+
+        hotel.removeAmenity(id_amenity);
+        return repository.save(hotel);
     }
 
-    public boolean deleteById(long id) {
-        try {
-            repository.deleteById(id);
-            return true;
-        } catch (Exception e) {
-            return false;
-        }
+    public void deleteById(long id) {
+        Optional<Hotel> optionalHotel = repository.findById(id);
+
+        if(!optionalHotel.isPresent())
+            throw new NotFoundException("Hotel with id: " + id + " not found");
+
+        repository.deleteById(id);
     }
 }
